Adds new functions for fishing friday.
Adds in the new 'fridayStatus', used for checking if the currently used server is enabled for Fishing Friday activities (which will be implemented later.) This saves the status to a json file, and is a simple functions for more advanced functions to check later.

Adds in '+timeUntilFriday', a function that takes the current time and returns the time until friday. Can be used at any point except for friday, and is for fun.

Applicable unit tests have been added, as well as some documentation.