Rewrites fishFactOperations.py and fishFactTests.py
This commit is a large commit that changes the content of fishFactOperations.py to better practices.
The function readFishFactData was added, and it takes a jsonFile to read, and returns a pandas dataframe.
The function saveFishFactData was added, and it takes a pandas dataframe and jsonfile and save it to the directory holding the python file.
Each of the previously existing functions must be given a fish fact dataframe to do their operations on, and will return the modified dataframe.
These should all work better in terms of coding practices, and make the program more consistent and easier to test.